mcm-snapbi-get-signature/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
python-dotenv==1.0.1
aiohttp==3.9.3
PyJWT==2.10.1
starlette==0.45.3